Fixed Indentation + TurtleTest1 Reward
Fixed a bug caused by an indentation in the learn function which led to a much higher LR than desired, which was a huge contributor to the instability problems faced in version 1.0.0. Also improved the accuracy of the algorithm with the TurtleTest1 environment from 0.5% to 11-12% by improving the reward function to take into account how long the agent had already been in the current episode.